refactor: move trading related API's to separate file
