+++ freqtrade/rpc/api_server/webserver.py
@@ -122,11 +122,12 @@ class ApiServer(RPCHandler):
         from freqtrade.rpc.api_server.api_download_data import router as api_download_data
         from freqtrade.rpc.api_server.api_pair_history import router as api_pair_history
         from freqtrade.rpc.api_server.api_pairlists import router as api_pairlists
+        from freqtrade.rpc.api_server.api_trading import router as api_trading
         from freqtrade.rpc.api_server.api_v1 import router as api_v1
         from freqtrade.rpc.api_server.api_v1 import router_public as api_v1_public
         from freqtrade.rpc.api_server.api_webserver import router as api_webserver
         from freqtrade.rpc.api_server.api_ws import router as ws_router
-        from freqtrade.rpc.api_server.deps import is_webserver_mode
+        from freqtrade.rpc.api_server.deps import is_trading_mode, is_webserver_mode
         from freqtrade.rpc.api_server.web_ui import router_ui
 
         app.include_router(api_v1_public, prefix="/api/v1")
@@ -136,6 +137,12 @@ class ApiServer(RPCHandler):
             api_v1,
             prefix="/api/v1",
             dependencies=[Depends(http_basic_or_jwt_token)],
+        )
+        app.include_router(
+            api_trading,
+            prefix="/api/v1",
+            tags=["trading"],
+            dependencies=[Depends(http_basic_or_jwt_token), Depends(is_trading_mode)],
         )
         app.include_router(
             api_webserver,
